Add wall and crate collision
The walls and crates now move and then check for collision, if it
collides with an object it moves back to the old location and returns
false so the player does not move either. The crate does not compare
the location to itself. A basic template is generated for the outer
walls and 4 placed crates